[2024-09-19] score overflow prevention
score 가 int 여서 점수가
21억점이 넘어가면 overflow가 발생하는 것을 자료형을 long으로 변경해서 해결함.

+++ Assets/0.Scripts/ScoreKeeper.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class ScoreKeeper : MonoBehaviour
 {
     public static ScoreKeeper instance;
-    public static int score { get; private set; }
+    public static long score { get; private set; }
     float lastEnemyKillTime;    //마지막으로 적을 죽인 시간
     int streakCount;            //적을 연속으로 죽인 횟수
     float streakExpiryTime = 1; //연속 킬이 유지되는 시간
@@ -31,7 +31,7 @@ public class ScoreKeeper : MonoBehaviour
         lastEnemyKillTime = Time.time;
 
         // 점수 계산: 기본 점수(5)에 스트릭 보너스를 추가 (2의 스트릭 카운트 제곱을 더함)
-        score += 5 + (int)Mathf.Pow(2, streakCount);
+        score += 5 + (long)Mathf.Pow(2, streakCount);
     }
 
     void OnPlayerDeath()
